Making Microarrays/Utilities/Extract samples from dataset less sensitive about phenodata column names.
When reading in the phenodata, by default R will check that column names are valid R variable names (e.g. cannot contain spaces or start with a number). As the user interface allows users to create illegal column names in phenodata (like containing spaces), this can break tools that have a METACOLUMN_SEL parameter, as the column names in the phenodata read into R will no longer be the same as was displayed in the user interface. This can be fixed by adding check.names=FALSE to the R command reading in the phenodata (of each and every tool), or alternatively by preventing creation of illegal column names in the phenodata in the user interface.